router structure
